More work on the settings component, actually looks like the settings panel now.
