+++ src/plugins/settings/components/plugins/layout.html
@@ -1,0 +1,1 @@
+<i>TODO</i>
