--- src/plugins/settings/layout.html
@@ -1,3 +1,0 @@
-<lol-uikit-dialog-frame class="ace-settings-container" orientation="bottom" frame="bordered">
-    <lol-uikit-flat-button @click="close()">DONE</lol-uikit-flat-button>
-</lol-uikit-dialog-frame>
