--- src/plugins/settings/style.styl
@@ -1,2 +1,0 @@
-.ace-settings-container
-    width 700px
